Notify WifiScanningService of scan failure upon iface deletion
Previously, WifiScanningService had no knowledge of iface
deletion, so if an iface was removed during a scan, it would
get stuck waiting for a scan result that never arrives.

Fixed to send a scan failure notification upon iface removal.
(Note: reportScanFailure() only sends a notification if a scan
was in progress, so this is safe even without an ongoing scan.)

Also converted an anonymous class to a lambda.

Bug: 178237359
Test: manual
Change-Id: Ic3a0b33dcd6189ccf6ee6f072a40b0d252abf90b
Merged-In: Ic3a0b33dcd6189ccf6ee6f072a40b0d252abf90b