ActiveModeWarden: Always propagate setRole() to ClientModeManager
Because of the IMS de-reg delay, the wifi toggle off takes a while to
get processed. So, ClientModeManager.getRole() can return the transient
state. Always propagate the role changes to ClientModeManager since it
is built to handle these. The recent changes in ag/13298767 added this
role check (mostly for unit test purposes). Fixed all the new tests
added in  ag/13298767 to account for the additional setRole calls.

Bug: 177924816
Test: atest com.android.server.wifi
Test: atest android.net.wifi.aware.cts.SingleDeviceTest
Change-Id: I5f5365b4e97485c42653103d42c63ede450e3a28