Allow the modern way to get ConnectivityManager
This removes a number of checks that no more interactions
were done on ConnectivityManager because this is
overspecified considering the reliance on third-party
libraries like NetworkFactory. This is exemplified in the
failures happening if these checks are not removed ;
methods like registerNetworkProvider used to be called
by NetworkFactory, but the old code would not see them
because it would not mock the instances of CM obtained
through the modern method.

Upcoming changes to NetworkFactory will also reshuffle
a little bit what hidden methods of ConnectivityManager
will call, so amending the changes to test for these
calls to the internals of ConnectivityManager is also
overspecifying.

Bug: None
Test: FrameworksWifiTests
Change-Id: I18683c28bf05928a0ce5ddaf44671385b8d9e7f0
(cherry picked from commit 7c9acd5aeab559059ea915c8944708f021f70edc)