WifiScoreReport: Send onStop to scorer on linger
When the previous primary CMM is forced to linger during MBB, send a
session end to external scorer (if registered). Even though we have not
disconnected from the previous network, logically the score is no longer
under the control of the external scorer. When the new primary CMM takes
over, the external scorer will get a new session start for the new
connection after MBB.

Bug: 175887153
Test: atest com.android.server.wifi
Test: atest android.net.wifi.cts.ConnectedNetworkScorerTest

Change-Id: Ib709b44949217b218838d34e823de63344af6888