Network insufficient if low rssi and no active traffic
The current logic of labelling a network as insufficient due to no
active traffic alone seems a bit too eager, as no active traffic does
not necessarily mean the network is unusable. Combining low rssi and the
active traffic check together also helps the device stay connected to
the same network during wifi calling when the rssi is low.

Also fixes an incorrectly mocked value in the unit test.

Bug: 175263854
Test: atest com.android.server.wifi
Change-Id: I7cbb466d69c79b1a376486d8a9548fa2e32d3905