ClientModeImpl: Transition to DisconnectedState on auth failure
When an auth failure is received, it marks the end of the connection
attempt as far as wifi framework is concerned. So, transition to
"DisconnectedState" on receiving this event (similar to assoc-reject,
network not found, etc). We are already reporting the connection attempt
end to various modules via reportConnectionAttemptEnd. However, if we
don't transition to "DisconnectedState", we will be invoking
reportConnectionAttemptEnd() again when NETWORK_DISCONNECTION_EVENT is
received from supplicant (auth failure is generally followed
NETWORK_DISCONNECTION_EVENT).

Also, removed an unused |networkId| variable and modified some unit
tests to verify this connection attempt end tracking.

Bug: 174974153
Test: atest ClientModeImplTest
Change-Id: Ibae9ac3c21185cad775465a31d13ed517e7b32e7