ActiveModeWarden: Change existing secondary CMM role
If there is already a secondary CMM which is created for connecting to
the same ssid/bssid as a newly requested one. Change the role of the
existing one & hand it out to the new client. The existing client of the
secondary CMM will be notified of this change via the ModeChange
callbacks (& should handle the necessary cleanup/recovery if necessary).

Bug: 158666312
Test: atest com.android.server.wifi
Change-Id: I4e289c808ae7bab18ad94c13a9c65e93feb61a62