ClientModeImpl: Always move to DisconnectedState on suppl state change
When we receive a "DISCONNECTED" state transition from supplicant,
always clear all the connection details and moved to DisconnectedState.
Removed the check for whether the notification is for a connecting
network. In this scenario, the notification is for a connected network.
However, this is a failsafe code path (we should get a
NETWORK_DISCONNECTION_EVENT in normal scenarios), so always clear out
connection details and start fresh.

Also,
i. synchronize the handling of this event in RoamingState & ConnectableState.
ii. Fix unit test static variable naming.

Bug: 179620232
Test: atest com.android.server.wifi
Change-Id: I5a616ac42bad44b62d80be5d71b139ceb79f179b