CMM: handle iface down when MAC randomization enabled and not connecting
When we are connecting to a network with MAC randomization,
we will set the iface down/up. But, we should only ignore
the iface down event while connecting. If we receive an
iface down event while not connecting, we should act on it,
regardless of whether MAC randomization is enabled.

Bug: 179906688
Test: atest FrameworksWifiTests
Change-Id: I70c924ec544f90a1597b0725e0685c936526bbea