Extract WifiNetworkAgent into its own class to facilitate unit testing
ClientModeImplTest uses INetworkAgent.aidl, which
is new in S and didn't exist in R. This causes
the test to fail when run on R. To fix this,
remove all references to INetworkAgent in
ClientModeImplTest.

Create a separate WifiNetworkAgent and callback
to decouple the Wifi-specific behavior from the
generic NetworkAgent behavior, and enables
mocking.

Bug: 134538181
Bug: 179512200
Test: atest FrameworksWifiTests (on S, R still has some failures to be fixed in later CLs)
Test: Boots up and connects to Wifi
Change-Id: I77c4ea0dd3d9b64bae59b0e1c3fe8e1275de20de