wifi: handle coex and wifi connected in SAP bridged mode
Before enabling SAP.
1. When wifi connect to unavailable channel, fallback to single AP.
2. When no available channel because coex, fallback to single AP.

After bridged AP enabled:
1. When wifi connect to unavailable channel
a. shutdown the instance which is running on the same band as STA.
b. shutdown the instance which is running on higher band.

2. When coex unsafe channel changed
a. shutdown the instance which no available channel on the current
running band,
b. shutdown the instance which is running on higher band.

Bug: 173999527
Bug: 162686273
Test: atest -c FrameworksWifiTests
Test: atest -c FrameworksWifiApiTests
Change-Id: I846a6a3b848dae7b7e0ffbacfff111093b15ae08