Non-primary CMMs are never explicitly connected
MBB is triggered by autojoin, so never explicitly
selected by the user. Same with other non-
primary roles.

isRecentlySelectedByTheUser doesn't ensure that
it never returns true for a non-primary network.

Consider the following (MBB-specific) scenario:

1. User manually tapped on network1
2. network1 went out of range
3. device switches to network2
4. network1 comes back in range
5. MBB switches back to network1

If this sequence of events all occur within 30
seconds, it is possible that a MBB network
could be treated as user selected.

If so, ConnectivityService would keep explicitly
selected networks up, even if they don't have
internet, which is undesirable for MBB.
Further, it could trigger the No Internet Dialog,
which allows the user to "Stay Connected" to a
network without internet. MBB does not currently
have a way to allow the user to stay connected
to a no internet network.

Thus, always mark non-primary STAs as not
explicitly connected.

Bug: 160346062
Test: atest ClientModeImplTest
Change-Id: I29c8ed8825587c2433ba1a5771e50cb30eb51844