WifiNetworkFactory: Use HAL API to disable fw roaming
Setting the bssid should have disabled roaming on a connection, but some
chips don't seem to be supporting this correctly. Use the explicit HAL
API to disable firmware roaming instead (if the firmware already
supports roaming disable on bssid set, then this is a no-op on those
devices).
Add an overlay to enable usage of this HAL API.

Bug: 178398607
Test: atest android.net.wifi.cts
Test: atest com.android.server.wifi
Change-Id: Id2f7e7aeeeeafb02c5d089b89a5cd45036b76621