ActiveModeWarden: Cleanup Hw multi STA API calls
Role switch to primary occurs last in the MBB sequence which results in
us in not configuring HW API's because of the current if check at the
start of |configureHwForMultiStaIfNecessary|.

Note: The previous logic works well for the other STA + STA use-cases
(because the secondary iface is created after the primary for all of
those use-cases), but does not work after the MBB switch.

Also, refactored some unit test helper methods.

Bug: 184109271
Test: atest com.android.server.wifi
Change-Id: I01ce592b999203ac1c5a1c4f3a542ab7dca11429