WifiShellCommand: process link probing in main thread
A recent change tightened up the packageName/UID
consistency check in AlarmManager. We are triggering
a link probe directly from a binder thread, which has
UID 0 since it is called from a rooted shell. The
context for WifiNl80211Manager is for the Wifi service,
which has packageName "android". Thus, they don't match.

Fix this by posting from binder thread to main Wifi
thread, which has the correct UID.

Bug: 185125428
Test: atest FrameworksWifiTests
Test: adb shell cmd wifi send-link-probe
Change-Id: I6221c619bc6f2730c3bd524a83be5bbcd218a7a8