Wifi: getUsableChannels error handling
The implementation is expected to follow the following rule:

If the HAL API returns SUCCESS - this implies supported and
information is available. An empty list means no available channels.

If the HAL API returns anything other than SUCCESS - it is either
unsupported or information not available. We must retun null which
would result in throwing UnsupportedOperationException exception.

Bug: 180755217
Test: atest android.net.wifi.cts.WifiManagerTest
Test: atest com.android.server.wifi.WifiVendorHalTest
Change-Id: I78f7679692d50d00802e95b88166edabcff41360