Update WifiInfoTest to check for immediate redaction.
Redaction in NetworkCapabilities is changing from redacting at
parcelling time to redacting immediately when makeCopy() is
called. This CL updates the WifiInfo tests to ensure that the
object is correctly redacted when makeCopy() is called.

The tests currently check the object against the redacted object
read back from a parcel, so they test both redaction and
parcelling. In order to maintain some test coverage for
parcelling, ensure the parcelled object is the same as the
redacted object using equals(). These test can only be run on
S and above because equals() was only introduced in S.

Bug: 183938194
Test: atest FrameworksWifiApiTests
Change-Id: I0109b4a05c312c48c9492f360d258a82bc009b58