Add the ability to match a band on a WiFi network.
• Add API to set a band on WifiNetworkSpecifier
• Make sure all WiFi networks have a WifiNetworkAgentSpecifier
• Compute the band from WifiInfo and store it in WNASpecifier
• Make sure the specifiers match each other correctly
• Add a measure of protection against matching the (B)SSID
  on regular WiFis

The intended use of this patch is for apps to be able to match
a WiFi network by its band. Also, while this is not supported
in stock Android, OEMs may replace parts of the stack to support
requesting a network by band, so that apps with appropriate
permissions can request the device connects to multiple bands
at the same time.

Bug: 181741503
Test: new test in cts/ConnectivityManagerTest
Change-Id: I1c2b023bcd53296c9aed221c74fcfb734aebdc6d