[WifiCoex] Lowest overlap channel must be lower than highest channel
When determining the lowest overlapped channel and highest overlapped
channel from an interfering frequency range, the lowest overlapped
channel must be lower than the highest overlapped channel. Generally
this is the case, but if the interference lands in the space in between
sequences of contiguous 5g channels (e.g. chan 68 to 96), then the
lowest overlap channel would be 96, and the highest overlap channel
would be 68. In this case, skip the calculations for this frequency
range since it does not land on any 5g channel.

Bug: 188404542
Test: atest CoexUtilsTest
Change-Id: Ia3339f8d9b2f19e788918a7622be856656c57c1b