Hide WifiNetworkSpecifier#redact.
ag/14572667 annotated this as @SystemApi, but did not update
system-current.txt accordingly. From an API perspective this
looks like it was actually a no-op because this method just
overrides the equivalent method in NetworkSpecifier, which is
already @SystemApi. However, it does break the CTS which checks
that @SystemApi annotations match the API text files.

Fix this by removing the @SystemApi annotation. This should not
impact any callers because nothing calls this method directly;
callers only call the redact() method on the superclass, whichis
already @SystemApi.

Bug: 188748612
Fix: 188748612
Test: builds, boots
Test: atest CtsSystemApiAnnotationTestCases android.signature.cts.api.AnnotationTest#testAnnotation
Change-Id: I0137e13c565aa831a1e1ae01720bf2c7c85cd863