Bug fix of bandwidth estimation
Bug fix of bandwidth estimation metrics: report AvgBandwidthKbps in wifi metrics based on the daily stats instead of the device historical average. That's because avgBandwidthKbps and count are meant to be daily stats which will be aggregated in the pipeline. Reporting the device historical average will result in over-counting of some stats.

In addition, the memory store of some devices may already have invalid values due to the early potential overflow bug. Reset them if invalid values are detected.

Bug: 189939948
Test: run mixed traffic for 30 mins and check reported BW
Test: atest com.android.server.wifi
Change-Id: I12794b17a80977a0982555da4388dbc19b5c29e6