wifi: Deep copy the object before send it
The binder (callback aidl) copy may not be executed when sender/receiver
are the same process.
For instance: In android GO, both of wifi stack and tethering stack are
in the same process. It will pass the instance directly without do
copying.
It will cause use the reference object directly in WifiManager.

Bug: 190782091
Test: atest -c FrameworksWifiTests
Test: Verified on target device per b/190782091#comment47
Change-Id: I56d1840e3b400e70e5698258b0ab5686ff46f0b1