Match mixed network environment correctly
In mixed security network environments, we need to filter out APs
with the stronger security type when the current saved network
supports the weaker security type, and the stronger security type
was added by auto-upgrade, and auto-upgrade feature is disabled.
This is needed for devices with the auto-upgrade feature disabled so
that only relevant networks with the weaker security are considered for
network selection.

Bug: 191857418
Test: atest WifiNetworkSelectorTest ScanResultMatchInfo
Test: Manual test with saved and suggested networks
Change-Id: Ia77286488ad7c9360a83dc25c8954c88e38d34cf