Only return WPA2 enterprise config if both WPA2 and WPA3 are available
R WifiTracker depends on the security type of each config with the
same SSID to be unique. Since R WifiTracker doesn't understand WPA3
enterprise, it will map it to the same security type as WPA2 enterprise
and crash. We should only return the WPA2 config in R if both types are
available.

Bug: 192558031
Test: atest FrameworksWifiTests
Change-Id: Ibc7576c16f27ac630717d2ea0a225ed8c3c78784