Handle S-on-R security types for DPP configurator network IDs
WifiServiceImpl#startDppAsConfiguratorInitiator takes an external
networkId as an argument. This networkId should be stripped of the
S-on-R security type that is added to networkId passed to external
clients, so that it may be referenced correctly internally. Likewise,
we should also add the security type to the networkId passed to
external clients through IDppCallback on enrollee success.

Bug: 193371401
Test: atest WifiServiceImplTest DppManagerTest android.net.wifi.cts.EasyConnectStatusCallbackTest
Change-Id: I307f0ce89e9ee7843c81c35597146b9bc7d2e8f5